chore/votePage
