+++ src/components/Calendar/CompleteButton.tsx
@@ -1,20 +1,18 @@
 import React, { useState } from "react";
 import * as S from "./CompleteButton.styles";
 import { Button } from "../Button";
-import { useNavigate, useLocation } from "react-router-dom";
 import { showToastWithMessage, isCompleteEnabled } from "./utils/Complete.utils";
 
 interface CompleteButtonProps {
   startDate: Date | null;
   endDate: Date | null;
   mode: "date" | "flexible";
+  onComplete: () => void;
 }
 
-const CompleteButton: React.FC<CompleteButtonProps> = ({ startDate, endDate, mode }) => {
+const CompleteButton: React.FC<CompleteButtonProps> = ({ startDate, endDate, mode, onComplete }) => {
   const [showToast, setShowToast] = useState(false);
   const [toastMessage, setToastMessage] = useState("");
-  const navigate = useNavigate();
-  const location = useLocation();
 
   const handleCompleteClick = () => {
     if (!startDate || !endDate) {
@@ -31,12 +29,7 @@ const CompleteButton: React.FC<CompleteButtonProps> = ({ startDate, endDate, mod
       return;
     }
 
-    // 현재 경로를 기준으로 이동 경로 설정
-    if (location.pathname.includes("/functional")) {
-      navigate(`/functional/vote`);
-    } else if (location.pathname.startsWith("/vote")) {
-      navigate(`/vote/success`);
-    }
+    onComplete();
   };
 
   return (
