+++ src/pages/Vote/ConfirmPage/_components/ConfirmFailCard/ConfirmFailCard.tsx
@@ -2,10 +2,10 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import * as S from "./Styles";
 import Card from "../../../../../components/Card";
-import { afterData } from "../../../VoteCompletePage/_data/afterData";
 import LinkIcon from "../../../../../assets/icons/LinkIcon.svg?react";
 import ConfirmMessage from "./ConfirmMessage";
 import { Button } from "../../../../../components/Button";
+import { voteData } from "../../../voteData";
 
 const ConfirmFailCard: React.FC = () => {
   const navigate = useNavigate();
@@ -27,7 +27,7 @@ const ConfirmFailCard: React.FC = () => {
       </Button>
 
       <Card.Image>  
-        <img src={afterData.imageSrc} alt="placeholder" 
+        <img src={voteData.imageSrc} alt="placeholder" 
         style={{width:"100%",height:"100%",borderRadius:"1.5rem",objectFit:"cover"}} />
       </Card.Image>
       
@@ -36,9 +36,9 @@ const ConfirmFailCard: React.FC = () => {
         <S.InfoItem>
           <S.LocationWrapper>
             <Card.Item label="장소">
-              <S.TruncatedText>{afterData.location}</S.TruncatedText>
+              <S.TruncatedText>{voteData.location}</S.TruncatedText>
             </Card.Item>
-            <S.IconWrapper onClick={() => handleCopyToClipboard(afterData.location)}>
+            <S.IconWrapper onClick={() => handleCopyToClipboard(voteData.location)}>
               <LinkIcon />
             </S.IconWrapper>
           </S.LocationWrapper>
@@ -48,14 +48,14 @@ const ConfirmFailCard: React.FC = () => {
 
         {/* 총액 */}
         <S.InfoItem>
-          <Card.Item label="총액">{afterData.total}</Card.Item>
+          <Card.Item label="총액">{voteData.total}</Card.Item>
         </S.InfoItem>
 
         <Card.Divider />
 
         {/* 기간 */}
         <S.InfoItem>
-          <Card.Item label="기간">{afterData.totalDuration}</Card.Item>
+          <Card.Item label="기간">{voteData.totalDuration}</Card.Item>
         </S.InfoItem>  
       </S.InfoContainer>
