+++ src/pages/Vote/Travel/index.tsx
@@ -1,5 +1,17 @@
 import TravelCard from "./_components/TravelCard";
 
 export default function OverView({
-  
-})
+    onNext
+}: {
+    onNext:()=>void
+}){
+
+    
+    return (
+        <>
+            <TravelCard
+                onNext={onNext}
+            />
+        </>
+    )
+}
