--- src/pages/Vote/VoteAgreePage/Vote.styles.ts
@@ -1,6 +1,0 @@
-
-
-
-
-
-
