--- src/pages/Vote/VoteAgreePage/VoteAgreePage.tsx
@@ -1,11 +1,0 @@
-// 여행 가보자고 - 투표 동의 화면
-import React from "react";
-import VoteCard from "./_components/VoteCard/VoteCard";
-
-const VoteAgreePage: React.FC = () => {
-    return (
-        <VoteCard />
-    );
-};
-
-export default VoteAgreePage;
