--- src/pages/Vote/VoteAgreePage/VotePage.tsx
@@ -1,12 +1,0 @@
-// VotePage.tsx
-import React from "react";
-import TravelCard from "./_components/TravelCard/TravelCard";
-
-const VotePage: React.FC = () => {
-    return (
-        <TravelCard />
-    );
-};
-
-export default VotePage;
-
