+++ src/pages/Vote/VoteCard/index.tsx
@@ -1,0 +1,18 @@
+import VoteCard from "../VoteCard";
+
+export default function OverView({
+  onAgree,
+  onDisagree,
+}: {
+  onAgree: () => void;
+  onDisagree: () => void;
+}) {
+  return (
+    <>
+        <VoteCard
+          onAgree={onAgree}
+          onDisagree={onDisagree}
+        />
+    </>
+  );
+}
