--- src/pages/Vote/VoteCompletePage/VoteResult.styles.ts
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledCard = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-export const CardWrapper = styled.div`
-  margin: 0.1rem 1.25rem;
-`;
-
-export const Content = styled.div`
-  font-size: 14px;
-  color: #6E6E6E;
-  text-align: center;
-  margin: 15px 0;
-`;
