--- src/pages/Vote/VoteCompletePage/_components/Duration.styles.ts
@@ -1,27 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: flex-start; 
-  align-items: flex-start; 
-`;
-
-export const InfoItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  color: #fff;
-`;
-
-export const InfoLabel = styled.div`
-  display: flex;
-  color: #fff;
-  font-weight: 500;
-  font-size: 12px;
-  margin-bottom: 3px;
-  padding: 5px;
-  align-items: flex-start;
-  justify-content: flex-start;
-`;
