--- src/pages/Vote/VoteCompletePage/_components/ResultCard/Styles.ts
@@ -1,48 +1,0 @@
-import styled from "styled-components";
-
-export const Image = styled.img`
-  width: 100%;
-  height: 100%;
-  border-radius: 1.5rem;
-  object-fit: cover;
-`;
-
-export const InfoItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start; 
-  justify-content: flex-start;
-  padding: 20px 0;
-  color: #fff;
-`;
-
-export const LocationWrapper = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-export const TruncatedText = styled.div`
-  display: flex;
-  flex-direction: column;
-  color: #fff;
-`;
-
-export const IconWrapper = styled.div`
-  cursor: pointer;
-  color: #007bff;
-
-  &:hover {
-    color: #0056b3;
-  }
-
-  svg {
-    width: 16px;
-    height: 16px;
-  }
-`;
-
-export const InfoContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start; 
-`;
