+++ src/pages/Vote/VoteDate/index.tsx
@@ -1,0 +1,9 @@
+import VoteDate from "../VoteDate";
+
+export default function OverView() {
+  return (
+    <>
+      <VoteDate />
+    </>
+  );
+}
