--- src/pages/Vote/VoteDatePage/Styles.ts
@@ -1,21 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  color: #fff;
-`;
-
-export const Typography = styled.div`
-  font-size:1rem;
-  color:#ffffff;
-  font-weight:500;
-  line-height:1.178rem;
-  position: relative;
-`
-export const HeaderContainer = styled.div`
-  position: relative;
-  top: 0;
-  left:0;
-  width: 100%;
-`;
