+++ src/pages/Vote/VotePage.tsx
@@ -1,16 +1,16 @@
 import React, { useState } from "react";
 import { useFunnel } from "../../hooks/useFunnel/useFunnel";
-import VoteDatePage from "./VoteDatePage/VoteDatePage";
 import ConfirmSuccessPage from "./ConfirmPage/ConfirmSuccessPage/ConfirmSuccessPage";
 import ConfirmFailPage from "./ConfirmPage/ConfirmFailPage/ConfirmFailPage";
 import * as S from "./_components/Vote.styles";
-import TravelCard from "./VoteAgreePage/_components/TravelCard/TravelCard";
-import VoteCard from "./VoteAgreePage/_components/VoteCard/VoteCard";
-import VoteResult from "./VoteCompletePage/VoteResult";
+import TravelCard from "./Travel/_components/TravelCard";
+import VoteCard from "./VoteCard/_components/VoteCard";
+import VoteDate from "./VoteDate/_components/VoteDate";
+import VoteResult from "./VoteResult/_components/VoteResult";
 
 const VotePage: React.FC = () => {
   const { Funnel, Step, setStep } = useFunnel("투표메인");
-  const [voteType, setVoteType] = useState<"찬성" | "반대" | null>(null); // 상태 추가
+  const [voteType, setVoteType] = useState<"찬성" | "반대" | null>(null); 
 
   return (
     <S.Container>
@@ -33,7 +33,7 @@ const VotePage: React.FC = () => {
         </Step>
 
         <Step name="날짜지정">
-          <VoteDatePage
+          <VoteDate
             onNext={() => {
               setStep("결과"); // 결과(찬성)로 이동
             }}
@@ -42,7 +42,7 @@ const VotePage: React.FC = () => {
 
         <Step name="결과"> 
           <VoteResult
-            type={voteType} // 찬성 or 반대
+            type={voteType!} // 찬성 or 반대
             onNext={() =>
               voteType === "찬성" 
                 ? setStep("찬성확인")
