+++ src/pages/Vote/VoteResult/_components/DurationInfo.tsx
@@ -1,14 +1,12 @@
 import React from "react";
 import Card from "../../../../components/Card";
-import * as S from "./Duration.styles";
-import { voteData } from "../../VoteAgreePage/_components/VoteCard/voteData";
+import * as S from "../../_components/Vote.styles"
+import { voteData } from "../../voteData";
 
 const DurationInfo: React.FC = () => {
   return (
     <S.Wrapper>
-      <S.InfoItem>
         <Card.Item label="기간">날짜지정 : {voteData.duration}</Card.Item>
-      </S.InfoItem>
     </S.Wrapper>
   );
 };
