+++ src/pages/Vote/VoteResult/_components/VoteComponent.styles.ts
@@ -5,7 +5,7 @@ export const Container = styled.div`
   padding: 0; 
   display: flex;
   flex-direction: column; 
-  gap: 9px;
+  gap: 0.563rem;
 `;
 
 export const VoteButton = styled.button<{ $isSelected: boolean; $selectedColor: string }>`
@@ -13,9 +13,9 @@ export const VoteButton = styled.button<{ $isSelected: boolean; $selectedColor: 
   justify-content: space-between;
   align-items: center;
   width: 100%;
-  padding: 12px 20px;
-  border-radius: 16px;
-  font-size: 14px;
+  padding: 1rem 1rem;
+  border-radius: 1rem;
+  font-size: 0.875rem;
   font-weight: 300;
   color: #fff;
   background-color: ${({ $isSelected, $selectedColor }) =>
@@ -39,11 +39,11 @@ export const VoteButton = styled.button<{ $isSelected: boolean; $selectedColor: 
       position: absolute;
       top: 0;
       left: 0;
-      width: 28%; 
+      width: 5.75rem; 
       height: 100%;
       background-color: #434343;
-      border-top-right-radius: 16px;
-      border-bottom-right-radius: 16px;
+      border-top-right-radius: 1rem;
+      border-bottom-right-radius: 1rem;
       z-index: 1;
     }
   `}
@@ -52,25 +52,25 @@ export const VoteButton = styled.button<{ $isSelected: boolean; $selectedColor: 
 export const Text = styled.div`
   position: relative;
   z-index: 2; 
-  font-size: 14px;
+  font-size: 0.875rem;
   font-weight: 300;
 `;
 
 export const TextWrapper = styled.div`
   display: flex;
   align-items: center;
-  gap: 15px; 
+  gap: 0.938rem; 
 `;
 
 export const VoteCounter = styled.span`
-  font-size: 14px;
+  font-size: 0.875rem;
   font-weight: normal;
-  margin-left: 8px;
+  margin-left: 0.5rem;
 `;
 
 export const VoteMessage = styled.div`
   display: flex;
-  font-size: 14px;
+  font-size: 1rem;
   font-weight: semi-bold;
   color: #fff;
 `;
