+++ src/pages/Vote/voteData.ts
